Refactor pytorch_utils
