+++ algorithmic_efficiency/interop_utils.py
@@ -1,0 +1,14 @@
+import jax.dlpack
+import torch
+
+from algorithmic_efficiency import spec
+
+
+def jax_to_pytorch(x: spec.Tensor, take_ownership: bool = False):
+  return torch.utils.dlpack.from_dlpack(
+      jax.dlpack.to_dlpack(x, take_ownership=take_ownership))
+
+
+def pytorch_to_jax(x: torch.Tensor):
+  x = x.contiguous()  # https://github.com/google/jax/issues/8082
+  return jax.dlpack.from_dlpack(torch.utils.dlpack.to_dlpack(x))
