+++ algorithmic_efficiency/workloads/wmt/wmt_pytorch/workload.py
@@ -11,9 +11,9 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 
 from algorithmic_efficiency import param_utils
 from algorithmic_efficiency import spec
-from algorithmic_efficiency.pytorch_utils import jax_to_pytorch
+from algorithmic_efficiency.interop_utils import jax_to_pytorch
+from algorithmic_efficiency.interop_utils import pytorch_to_jax
 from algorithmic_efficiency.pytorch_utils import pytorch_setup
-from algorithmic_efficiency.pytorch_utils import pytorch_to_jax
 from algorithmic_efficiency.workloads.wmt import bleu
 from algorithmic_efficiency.workloads.wmt import decode
 from algorithmic_efficiency.workloads.wmt.wmt_pytorch.models import Transformer
